# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTodoRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  dto.TodoResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      priority:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.UpdateTodoRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      status:
        enum:
        - not_started
        - in_progress
        - completed
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    type: object
  dto.UpdateTodoStatusRequest:
    properties:
      status:
        enum:
        - not_started
        - in_progress
        - completed
        type: string
    required:
    - status
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad request
        type: string
      request_id:
        example: abc-123
        type: string
    type: object
  response.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Success
        type: string
      pagination:
        $ref: '#/definitions/response.Pagination'
      request_id:
        example: abc-123
        type: string
    type: object
  response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Success
        type: string
      request_id:
        example: abc-123
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A high-performance todo list microservice built with Go and Clean Architecture
    principles.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Todo List API
  version: "1.0"
paths:
...
  /todos:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of todos for the authenticated user with
        optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - not_started
        - in_progress
        - completed
        in: query
        name: status
        type: string
      - description: Filter by priority
        enum:
        - low
        - medium
        - high
        in: query
        name: priority
        type: string
      - description: Search in title and description
        in: query
        maxLength: 100
        name: search
        type: string
      - description: Filter todos due after this date (RFC3339 format)
        format: date-time
        in: query
        name: due_date_from
        type: string
      - description: Filter todos due before this date (RFC3339 format)
        format: date-time
        in: query
        name: due_date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Todos retrieved successfully
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "400":
          description: Invalid user ID or request format
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: List todos
      tags:
      - Todos
    post:
      consumes:
      - application/json
      description: Create a new todo item for the authenticated user
      parameters:
      - description: Todo details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTodoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Todo created successfully
          schema:
            $ref: '#/definitions/dto.TodoResponse'
        "400":
          description: Invalid request format or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new todo
      tags:
      - Todos
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a todo item by its ID (only own todos)
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Todo deleted successfully
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid todo ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Todo not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a todo
      tags:
      - Todos
    get:
      consumes:
      - application/json
      description: Retrieve a specific todo item by its ID (only own todos)
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Todo retrieved successfully
          schema:
            $ref: '#/definitions/dto.TodoResponse'
        "400":
          description: Invalid todo ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Todo not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get a todo by ID
      tags:
      - Todos
    put:
      consumes:
      - application/json
      description: Update an existing todo item by its ID (only own todos)
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated todo details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Todo updated successfully
          schema:
            $ref: '#/definitions/dto.TodoResponse'
        "400":
          description: Invalid request format or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Todo not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update a todo
      tags:
      - Todos
  /todos/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a todo item by its ID (only own todos)
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTodoStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Todo status updated successfully
          schema:
            $ref: '#/definitions/dto.TodoResponse'
        "400":
          description: Invalid request format or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Todo not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update todo status
      tags:
      - Todos
...
securityDefinitions:
  Bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"